pyproject.toml
[project]
name = "miyori"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "SimonPurdie", email = "hoopmonkey@gmail.com" }
]
requires-python = ">=3.11, <3.13"
dependencies = [
    "ddgs>=9.10.0",
    "google-genai>=1.56.0",
    "hdbscan>=0.8.41",
    "kokoro>=0.9.4",
    "numpy>=2.4.0",
    "pip>=25.3",
    "pvcobra==2.1.0",
    "pvporcupine>=4.0.1",
    "pyaudio>=0.2.14",
    "scikit-learn>=1.8.0",
    "sounddevice>=0.5.3",
    "vosk>=0.3.45",
    "winshell>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/miyori"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]

[project.scripts]
miyori = "miyori.main:main"

